changes to page and components

+++ api/controllers/user.controller.js
@@ -1,33 +1,39 @@
 import { errorHandler } from "../utils/error.js";
-import User from "../models/user.model.js";
+import User from "../models/user.models.js";
 import bcryptjs from "bcryptjs";
 
-export const logout = (req, res) => {
-  try {
-    res
-      .clearCookie("access_token")
-      .status(200)
-      .json("User has been logged out");
-  } catch (error) {
-    next(error);
-  }
+export const test = (req, res) => {
+  res.json({ message: "API is working!!" });
 };
 
 export const updateUser = async (req, res, next) => {
-  if (req.user.userId !== req.params.userId) {
+  if (req.user.id !== req.params.userId) {
     return next(
       errorHandler(403, "You are not authorised to update this user")
     );
   }
   if (req.body.password) {
     if (req.body.password.length < 6) {
-      return next(errorHandler(400, "Password must be at least 6 characters"));
+      return next(errorHandler(400, "Passowrd must be at least 6 characters"));
     }
     req.body.password = bcryptjs.hashSync(req.body.password, 10);
   }
-  if (req.body.fullname) {
-    if (!req.body.fullname.match(/^[a-zA-Z ]+$/)) {
-      return next(errorHandler(400, "Your name can only contain letters"));
+  if (req.body.username) {
+    if (req.body.username.length < 6 || req.body.username.length > 20) {
+      return next(
+        errorHandler(400, "Username must be between 7 and 20 characters")
+      );
+    }
+    if (req.body.username.includes(" ")) {
+      return next(errorHandler(400, "Username cannot contain spaces"));
+    }
+    if (req.body.username !== req.body.username.toLowerCase()) {
+      return next(errorHandler(400, "Username must be lowercase"));
+    }
+    if (!req.body.username.match(/^[a-zA-Z0-9]+$/)) {
+      return next(
+        errorHandler(400, "Username can only contain letters and numbers")
+      );
     }
   }
   try {
@@ -36,9 +42,11 @@ export const updateUser = async (req, res, next) => {
       {
         $set: {
           fullname: req.body.fullname,
+          username: req.body.username,
           email: req.body.email,
+          profilePicture: req.body.profilePicture,
+          authorbio: req.body.authorbio,
           password: req.body.password,
-          avatar: req.body.avatar,
         },
       },
       { new: true }
@@ -51,12 +59,23 @@ export const updateUser = async (req, res, next) => {
 };
 
 export const deleteUser = async (req, res, next) => {
-  // if (!req.user.isAdmin && req.user.userId !== req.params.userId) {
-  //   return next(errorHandler(403, "You are not allowed to delete this user"));
-  // }
+  if (!req.user.isAdmin && req.user.id !== req.params.userId) {
+    return next(errorHandler(403, "You are not allowed to delete this user"));
+  }
   try {
     await User.findByIdAndDelete(req.params.userId);
     res.status(200).json("User has been deleted");
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const signout = (req, res, next) => {
+  try {
+    res
+      .clearCookie("access_token")
+      .status(200)
+      .json("User has been signed out");
   } catch (error) {
     next(error);
   }
